DCC-641 change ints to longs to prevent byte offset overflows
Former-commit-id: 56adc8dc620d0d0753ec97a4d637001bdc23bf95
Former-commit-id: f36a4f072adf7fab42ee94adba07bbc92b1f122a

+++ server/src/main/java/org/icgc/dcc/validation/cascading/StructuralCheckFunction.java
@@ -87,7 +87,7 @@ public class StructuralCheckFunction extends BaseOperation implements Function {
 
     TupleEntry arguments = functionCall.getArguments();
 
-    int offset = functionCall.getArguments().getInteger(ValidationFields.OFFSET_FIELD_NAME);
+    long offset = functionCall.getArguments().getLong(ValidationFields.OFFSET_FIELD_NAME);
     TupleState tupleState = new TupleState(offset);
 
     String line = arguments.getString(StructuralCheckFunction.LINE_FIELD_NAME);
